Web App/requirements.txt
argcomplete==1.12.1
blivet==3.3.1
boto3==1.16.21
botocore==1.19.21
certifi==2020.11.8
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
durationpy==0.5
Flask==1.1.2
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pip-tools==5.3.1
placebo==0.9.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pyudev==0.22.0
PyYAML==5.3.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.52.0
troposphere==2.6.3
urllib3==1.26.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.0
